Add files via upload
hardware oriented design where a extra datahandler was added to turn the word addr hex command into little endian byte addr command. Instruction mem is kept as a seperate portion and remained as close to the RISV-V original design as possible, and a top module is used to intergrate them. This way it is more hardware-wise reasonable